% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotprcurve.R
\name{plot.pr.curve}
\alias{plot.pr.curve}
\title{plot.pr.curve}
\usage{
\method{plot}{pr.curve}(predictions, title.text)
}
\arguments{
\item{text}{}
}
\value{

}
\description{
: auxilia curva roc
}
\examples{


}
