% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotroccurve.R
\name{plot.roc.curve}
\alias{plot.roc.curve}
\title{plot.roc.curve}
\usage{
\method{plot}{roc.curve}(predictions, title.text)
}
\arguments{
\item{text}{}
}
\value{

}
\description{
: auxilia curva roc
}
\examples{


}
