% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao_vari.R
\name{run.feature.selection}
\alias{run.feature.selection}
\title{Função para seleção de variáveis}
\usage{
run.feature.selection(num.iters = 20, feature.vars, class.var)
}
\arguments{
\item{variaveis}{}
}
\value{

}
\description{
Função para seleção de variáveis
}
