% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varivel_fator.R
\name{to.factors}
\alias{to.factors}
\title{Transformar a variavel em fator}
\usage{
to.factors(df, variables)
}
\arguments{
\item{df}{data.frame em analise}

\item{variavel}{nome da coluna entre (" ")}
}
\value{
variaveis tipo fator
}
\description{
Transformar a variavel em fator
}
