% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_contagem_NA's.R
\name{contagem_nas}
\alias{contagem_nas}
\title{Contagem de NA's
A funcao retorna o numero de NAs dentro de um vetor do data.frame}
\usage{
contagem_nas(df, variavel)
}
\arguments{
\item{variavel}{nome da coluna entre (" ")}

\item{base_dados}{data.frame em analise}
}
\value{
Quantidade de NAs
}
\description{
: usethis::use_pipe()
}
